pyproject.toml
[project]
name = "paper-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "hydra-core>=1.3.2",
    "marker-pdf>=1.7.1",
    "rich>=10.2.0",
    "sentence-transformers>=4.1.0",
    "torch==2.6.0",
    "torchvision>=0.21.0",
]
